Refactor server.py into more specific files under pages/.
